+++ app/src/main/java/org/mozilla/fenix/HomeActivity.kt
@@ -53,6 +53,7 @@ open class HomeActivity : AppCompatActivity() {
 
         setTheme(themeManager.currentTheme)
         DefaultThemeManager.applyStatusBarTheme(window, themeManager, this)
+        components.core.setEnginePreferredColorScheme()
 
         browsingModeManager = DefaultBrowsingModeManager(this)