+++ CHANGELOG.md
@@ -14,5 +14,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - #1049 - Add style for progress bar with gradient drawable
 - #1165 - Added doorhanger to the toolbar
 - #1195 - Adds telemetry for quick action sheet
+- #627 - Sets engine preferred color scheme based on light/dark theme
 ### Changed
 ### Removed